#43 : 충돌 방지를 위한 커밋

+++ src/main/java/com/ssg/dsilbackend/repository/PointManageRepository.java
@@ -3,5 +3,5 @@ package com.ssg.dsilbackend.repository;
 import com.ssg.dsilbackend.domain.Point;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PointManageRepository extends JpaRepository<Point, Integer> {
+public interface PointManageRepository extends JpaRepository<Point, Long> {
 }
