+++ src/main/java/com/ssg/dsilbackend/repository/ReplyManageRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ReplyManageRepository extends JpaRepository<Reply, Integer> {
+public interface ReplyManageRepository extends JpaRepository<Reply, Long> {
 
 }
